New version number  0.05
All "true", and "false" are now "TRUE", and "FALSE" due to new standards in devkit, and libogc. Vector is now guVector, and in gx_mesh.c the array 
used is no longer the enumerator of q_boolean, but now and array of ints to fix a crashing bug, and rendering issues. This appears to be a working 
version with the latest version of devkitpro, and libogc.